+++ nvi-evaluator/src/main/java/no/sikt/nva/nvi/evaluator/exceptions/NotACandidateException.java
@@ -1,0 +1,5 @@
+package no.sikt.nva.nvi.evaluator.exceptions;
+
+public class NotACandidateException extends RuntimeException {
+
+}
